### 2020-03-01 #2126 @whedon
[PRE REVIEW]: Omorfi: Open morphology of Finnish
**Submitting author:** @flammie (<a href="http://orcid.org/0000-0003-1207-5395">Tommi Pirinen</a>)
**Repository:** <a href="https://github.com/flammie/omorfi/" target ="_blank">https://github.com/flammie/omorfi/</a>
**Version:** 2020-05-11
