### 2020-03-01 #2126 @labarba
👋 @lorenanicole — This submission is in computational linguistics and natural language processing, and I think you are our only expert in these areas. Can you take it as handling editor? Thanks!
