### 2020-03-18 #2126 @arfon
:wave: @hugoledoux - I realize this isn't close to your fields/disciplines but given the pause in submissions for JOSS right now, this is the only paper currently without an editor assigned.

Are you OK for me to go ahead and assign this to you?
