### 2020-04-15 #2126 @flammie
no worries, we're all busy with random things at the moment. If we can recruit freely I'd probably suggest from the reviewer pool of nlp oss workshop: https://nlposs.github.io/2020/index.html#programme-committee. I personally only know alvations (Liling Tan) who is organising but he may be quite busy.
