### 2020-05-25 #2184 @hugoledoux
oh, I should have RTFD. Actually I did, just spotted the command in red, and moved on.

It's fixed now.
