### 2020-05-25 #2184 @alvations
Okay, it's a little overwhelming but definitely not tonight but I'll work on the review over the coming week nights.

A few questions on the JOSS review, should the comments on the review come here in this issue or should it go directly to https://github.com/flammie/omorfi/issues/new so that the problems in the review are corrected?

Also, regarding the high-level summary (the read-able paper), how do I get access to it for the review of that too?
