### 2020-05-25 #2184 @hugoledoux
yes it is at first.

You should indeed create issues in the original repo, and link to them here.

The idea is simply to verify/check that all the items in the checklist above (watch out, there's one for you and one for the other reviewer) are okay.

The PDF is available a few comments above here, or [here](https://github.com/openjournals/joss-papers/blob/joss.02184/joss.02184/10.21105.joss.02184.pdf)
