### 2020-05-25 flammie/omorfi#68 @alvations
Please add a plain-text LICENSE file 
A plain text `LICENSE` file needs to be included in the repository.

> **License**: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

P/S: As per review requested on https://github.com/openjournals/joss-reviews/issues/2184
