### 2020-05-25 #2184 @alvations
Thanks @hugoledoux for the clarifications!

Commenting out the partial reviews first, will undo the comments at the end.
