### 2020-05-25 flammie/omorfi#69 @alvations
On Community Guidelines
The current README.md is a little cluttered with many information. Maybe moving the information for OSS contribution should have been moved to a CONTRIBUTING.md file and still keep the `# Community Guidelines` header in the README but points it to the CONTRIBUTING file?

Also, regarding the `community guidelines` requirements for JOSS:

>  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

- `1)` can be moved into a separate `CONTRIBUTING.md` file as suggested
- `2)` can be pointing to the `Troubleshooting` section and https://flammie.github.io/omorfi/pages/troubleshooting.html as well as the github issue tracker
- `3)` there's a need for a contact point to exist for seeking support, perhaps be a little more explicit if support/questions are welcomed on Github issues, otherwise, pointer to a mailing list or email support would be good.

P/S: As per review requested on openjournals/joss-reviews#2184
