### 2020-05-27 19c68c5
show GNU GPL copying under LICENSE
fixes #68 (if link is ok?)
