### 2020-05-27 flammie/omorfi#68 @flammie
thanks for the report, I had GNU standard COPYING, I've made link for now, to see if it shows up fine...
