### 2020-05-27 50ab9d1
split part of readme to contributing doc
towards cleanups mentioned in #69
