### 2020-06-11 #2184 @desilinguist
Sorry for the delay. I just started the review but I can't proceed because I am finding the installation quite confusing. It's not clear to me what instructions should be followed.
