### 2020-06-12 #2184 @flammie
Yeah, there's no pip package or similar yet, but I can look into making one in near future. The minimum installation is to get dependencies which are packaged and you probably need to run autoreconf and configure to have helper scripts use right program paths. The language models can then be downloaded from latest release instead of compiling anew. The readme improvements are tracked in an issue in the repo, I'll try to clarify the instructions.
