### 2020-07-14 e014abd
Reorganise docs, fixes #69
