+++ README.md
@@ -1,36 +1,8 @@
 # Omorfi–Open morphology of Finnish
 
 This is a free/libre open source morphology of Finnish: a database, tools and
-APIs. This package is licenced under [GNU GPL version
-3](https://gnu.org/licenses/gpl.html), but *not necessarily later*. Licence can
-also be found in the [COPYING]() file in the root directory of this package.
-Other licences are possible by *all* the authors named in the [AUTHORS]() file.
-
-Omorfi has been used for a number of tasks:
-
-* morphological analysis
-* morphological segmentation
-* spell-checking and correction
-* information retrieval
-* ontologies
-* statistical machine translation
-* rule-based machine translation
-* language modeling
-* tokenisation and sentence boundary detection
-* stemming, lemmatisation and shallow morph analysis
-
-The lexical data of omorfi has been acquired from various sources with
-different original licences.  The dictionaries used in omorfi are [Nykysuomen
-sanalista](//kaino.kotus.fi) (LGPL),
-[Joukahainen](//joukahainen.lokalisointi.org) (GPL) and
-[FinnWordNet](//www.ling.helsinki.fi/research/finnwordnet) (Princeton
-Wordnet licence / GPL; relicenced with kind permission from University of
-Helsinki), [Finnish Wiktionary](//fi.Wiktionary.org) and [English
-Wiktionary](//en.wiktionary.org) (Creative Commons
-Attribution–ShareAlike). Some words have also been collected by omorfi
-developers and contributors and are GPLv3 like the rest of the package.
-
-These are the obligatory stamps of the day:
+APIs. Everything you need to build NLP applications processing Finnish language
+words and texts.
 
 [![Build Status](https://travis-ci.org/flammie/omorfi.svg?branch=develop)](https://travis-ci.org/flammie/omorfi)
 (stable master branch:
@@ -40,185 +12,91 @@ These are the obligatory stamps of the day:
 
 ## Documentation
 
-There is a [github-pages site for omorfi](//flammie.github.io/omorfi/), that
-contains both automatically generated information about omorfi and larger
-documentation, including some design rationale and historical remarks.
+For more detailed information, see [github pages for
+omorfi](//flammie.github.io/omorfi/).
 
-### Academic works
+## Citing and academic works
 
-Some technical background is detailed also in academic papers, these are
-collected on the [articles page on
-gh-pages](//flammie.github.io/omorfi/pages/Articles.html), including
-citation information.
+Citation information can be found in file
+[CITATION](https://github.com/flammie/omorfi/blob/develop/CITATION). For further
+details, see [omorfi articles](//flammie.github.io/omorfi/pages/Articles.html).
 
 ## Downloading and further information
 
-→ See also: [Release policy](https://flammie.github.io/omorfi/pages/Releases.html)
+Omorfi packages can be downloaded from github:
 
-Omorfi is currently hosted at github. [Omorfi's github
-repository](https://github.com/flammie/omorfi) contain most of the important
-information about omorfi: version control system for source codes, bug tracker
-for reporting bugs, and the stable releases as convenient packages.
+* [omorfi releases](https://github.com/flammie/omorfi/releases)
+
+or the most current version using git. For more information, see
+[Release policy](https://flammie.github.io/omorfi/pages/Releases.html)
 
 ## Dependencies
 
-*Before you start:* Apertium wiki has installation information for most
-dependencies on their [packaging-based installation instructions for
-Linux and WSL](http://wiki.apertium.org/wiki/Install_Apertium_core_using_packaging),
[…]
