+++ .github/ISSUE_TEMPLATE/word-change-request.md
@@ -1,0 +1,22 @@
+---
+name: Word change request
+about: Template for adding or removing words
+title: 'Word change request: [WORD]'
+labels: word
+assignees: ''
+
+---
+
+The following words are missing from omorfi:
+
+* [WORD]
+  * [URL that shows attested real world usage],
+  * [EXAMPLE usage in context]
+  * [POS part of speech],
+  * [inflectional PARADIGM], 
+* [copy-paste as many times as needed]
+
+Following word-items should be deleted from omorfi:
+
+* [WORD], [POS], [HOMONYM ID]
+  * explanation
