+++ src/python/README.md
@@ -1,0 +1,3 @@
+# omorfi-python
+
+Python API for omorfi (the open morphology of Finnish).
