### 2020-07-17 19c3014
add scripts to pip packaging
make python download script aware of current version for default
download target
