+++ src/python/setup.py
@@ -18,6 +18,21 @@ setuptools.setup(
         "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
         "Operating System :: OS Independent",
     ],
+    keywords="finnish nlp morphology",
     python_requires='>=3.5',
+    scripts=['omorfi-download.py',
+             'omorfi-conllu.py',
+             'omorfi-convert.py',
+             'omorfi-disamparsulate.py',
+             'omorfi-download.py',
+             'omorfi-factorise.py',
+             'omorfi-freq-evals.py',
+             'omorfi-ftb3.py',
+             'omorfi-segment.py',
+             'omorfi-sigmorphons.py',
+             'omorfi-tokenise.py',
+             'omorfi-unimorph.py',
+             'omorfi-vislcg.py',
+             'omorfi-wikitable.py'],
     install_requires=['hfst'],
 )
