+++ docs/pages/omorfi-spell(1).md
@@ -1,45 +1,0 @@
----
-layout: default
-title: "omorfi-spell(1)"
-category: man
-date: 2016-02-08 18:23:58
----
-
-
-```
-OMORFI-SPELL(1)                  User Commands                  OMORFI-SPELL(1)
-
-
-
-NAME
-       omorfi-spell - Spell-check Finnish words using omorfi
-
-SYNOPSIS
-       omorfi-spell.sh [OPTION] [FILENAME...]
-
-DESCRIPTION
-       Spell checks and corrects list of words.
-
-       -h, --help
-              Print this help dialog
-
-       -V, --version
-              Print version info
-
-       -v, --verbose
-              Print verbosely while processing
-
-       If no FILENAMEs are given, input is read from standard input.  This tool
-       uses hfst-ospell.
-
-COPYRIGHT
-       Copyright © 2015 Omorfi contributors.  Licence GPLv3: GNU GPL version  3
-       <http://gnu.org/licenses/gpl.html>
-       This  is  free  software:  you  are  free to change and redistribute it.
-       There is NO WARRANTY, to the extent permitted by law.
-
-
-
-OMORFI                             March 2015                   OMORFI-SPELL(1)
-```
-
