+++ docs/pages/omorfi-generate(1).md
@@ -1,48 +1,0 @@
----
-layout: default
-title: "omorfi-generate(1)"
-category: man
-date: 2016-02-08 18:23:58
----
-
-
-```
-OMORFI-GENERATE(1)               User Commands               OMORFI-GENERATE(1)
-
-
-
-NAME
-       omorfi-generate - generate Finnish word-forms from omor style specs
-
-SYNOPSIS
-       omorfi-generate.sh [OPTION] [FILENAME...]
-
-DESCRIPTION
-       Generate word-forms from line separated omorfi analyses.
-
-       -h, --help
-              Print this help dialog
-
-       -V, --version
-              Print version info
-
-       -v, --verbose
-              Print verbosely while processing
-
-       If no FILENAMEs are given, input is read from standard input.
-
-EXAMPLES
-       Generate strings from command-line:
-              omorfi-generate.sh
-
-COPYRIGHT
-       Copyright  ©  2015  Omorfi contributors Licence GPLv3: GNU GPL version 3
-       <http://gnu.org/licenses/gpl.html>
-       This is free software: you are  free  to  change  and  redistribute  it.
-       There is NO WARRANTY, to the extent permitted by law.
-
-
-
-OMORFI                             March 2015                OMORFI-GENERATE(1)
-```
-
