+++ docs/pages/TODO.md
@@ -1,13 +1,0 @@
----
-layout: default
-title: "TODO"
----
-
-
-# Introduction #
-
-Known bugs and limitations in omorfi.
-
-* ...
-
-See also https://github.com/flammie/omorfi/issues
