+++ docs/pages/CoveragesTop100Deltas.md
@@ -1,4588 +1,0 @@
----
-layout: default
-title: "CoveragesTop100Deltas"
-category: stats
-date: 2016-02-08 18:23:58
----
-
-
-# Coverage deltas
-
-Use this listings to track regressions. It is automatically generated and I
-merely >> the new data at the end. See:
-
-* [2013-03-19](#2013-03-19)
-* [2013-04-17](#2013-04-17)
-* [2014-10-14](#2014-10-14)
-* [2015-03-26](#2015-03-19)
-* [2015-09-04](#2015-09-04)
-
-# 2013-03-19 #
-
-## fi-gutenberg ##
-
-### Missing word forms ###
-
-| **Frequency** | **Word form** |
-|:--------------|:--------------|
-| 854 |  ett |
-| 715 |  Mathieu |
-| 604 |  sitte |
-| 273 |  Tressilian |
-| 247 |  Varney |
-| 241 |  Constance |
-| 229 |  Servosse |
-| 222 |  Jaampa |
-| 201 |  Lily |
-| 200 |  kauvan |
-| 172 |  Morange |
-| 165 |  nämät |
-| 165 |  Mathieun |
-| 154 |  aivankuin |
-| 153 |  FALSTAFF |
-| 151 |  HOMSANTUU |
-| 149 |  Beauchêne |
-| 144 |  taaskin |
-| 143 |  Laagje |
-| 139 |  huoneesen |
-| 136 |  Mellet |
-| 129 |  ell |
-| 119 |  Séguin |
-| 113 |  Bathseba |
-| 109 |  Lambourne |
-| 106 |  Sérafine |
-| 101 |  ennenkin |
-| 100 |  itsekin |
-| 99 |  vaikk |
-| 92 |  Margreta |
-| 92 |  eteensä |
-| 87 |  Aarnold |
-| 85 |  Sitte |
-| 81 |  Warrington |
-| 81 |  olis |
-| 80 |  Comfort |
-| 79 |  ia |
-| 78 |  kiini |
-| 77 |  iin |
-| 76 |  ainoatakaan |
-| 75 |  Mr |
-| 75 |  frankia |
-| 74 |  OF |
-| 72 |  THE |
-| 72 |  PROJECT |
-| 72 |  GUTENBERG |
-| 72 |  Ett |
-| 72 |  EBOOK |
-| 72 |  Blaise |
-| 72 |  Beidermann |
-| 71 |  kylläksi |
-| 71 |  issa |
-| 70 |  yht'äkkiä |
-| 67 |  Kyllähän |
-| 67 |  kehoitti |
-| 67 |  katteini |
-| 67 |  Constancen |
-| 66 |  GLOSTER |
-| 65 |  jälellä |
-| 64 |  Walters |
-| 64 |  Couteauska |
-| 64 |  Burleson |
-| 63 |  nuot |
-| 62 |  Tressilianin |
-| 62 |  kentiesi |
-| 61 |  Boutan |
-| 60 |  Wayland |
-| 60 |  Miks |
-| 60 |  Beauchênen |
-| 59 |  Sussexin |
-| 58 |  Ambroise |
-| 57 |  Nämät |
-| 56 |  Laagjen |
-| 56 |  itsekään |
-| 55 |  Verdenton |
-| 54 |  VÖRSKY |
-| 54 |  Pieter |
-| 54 |  Norinen |
-| 54 |  Metta |
-| 53 |  Amy |
-| 52 |  viimeinkin |
-| 52 |  ROSAMUNDA |
-| 52 |  Reine |
-| 52 |  ensimäinen |
-| 51 |  Jaampan |
[…]

[…]
