### 2020-07-28 f78c566
fix versions of some corpora
also zcat / gzcat mac os x fails
