### 2020-07-28 28aaf25
remove extra stuff
because github didnt support too many files
