### 2020-07-28 05feab0
generate new paradigms
