+++ docs/generate-paradigms.py
@@ -47,9 +47,6 @@ def main():
     ap.add_argument("--output", "-o", action="store", required=True,
                     type=argparse.FileType('w'),
                     metavar="OFILE", help="write docs OFILE")
-    ap.add_argument("--outdir", "-O", action="store", required=True,
-                    metavar="ODIR", help="write individual paradigms to " +
-                    "ODIR/paradigm.md")
     ap.add_argument("--fields", "-F", action="store", default=2,
                     metavar="N", help="read N fields from master")
     ap.add_argument("--separator", action="store", default="\t",
@@ -88,20 +85,17 @@ def main():
     print(file=args.output)
     print("* The _Notes_ column is informal hints for lexicographers to " +
           "discriminate the classes, for further details each paradigm has " +
-          "their own page with details and exampels linked ", file=args.output)
+          "their own page with details and exampels linked", file=args.output)
     print("* The _KOTUS_ column is official dictionary class, if you are  " +
-          "familiar with Finnish dictionaries. ", file=args.output)
+          "familiar with Finnish dictionaries.", file=args.output)
     print("* The _Harmony_ column is vowel harmony; back, front, both or N/A",
           file=args.output)
-    print("* The _Inflection tables_ column links to inflection tables of the" +
-          " example word, *full* for full-form list and *short* for " +
-          "wiktionary-style inflection tables.", file=args.output)
     # read from csv files
     print(file=args.output)
-    print("| **Paradigm** | **UPOS** | _Notes_ | KOTUS | Harmony | Infl. " +
-          "Tables |", file=args.output)
-    print("|:-------------|---------:|:--------|------:|:-------:|:------" +
-          "-------|", file=args.output)
+    print("| **Paradigm** | **UPOS** | _Notes_ | KOTUS | Harmony |",
+          file=args.output)
+    print("|:-------------|---------:|:--------|------:|:-------:|",
+          file=args.output)
 
     paradata = dict()
     with open(args.paradigms) as tsv_file:
@@ -128,18 +122,6 @@ def main():
                           "skipping following line completely:", file=stderr)
                     print(tsv_parts, file=stderr)
                     continue
-                outfilename = args.outdir + \
-                    tsv_parts['new_para'].replace('?', '_').replace('_', '/',
-                                                                    1)
-                outfile = open(outfilename + '.markdown', 'w')
-                print('---', file=outfile)
-                print('layout: paradigm', file=outfile)
-                print('paradigm:', tsv_parts['new_para'], file=outfile)
-                print('---', file=outfile)
-                print("### `", tsv_parts['new_para'], "`",
-                      file=outfile)
-                print(file=outfile)
-                print('* _' + tsv_parts['doc'] + '_', file=outfile)
                 if tsv_parts['new_para'] in paradata:
                     if not tsv_parts['doc']:
                         print("missing DOC:", tsv_parts['new_para'])
@@ -160,73 +142,16 @@ def main():
                         harmony = 'N/A'
                     else:
                         harmony = paradata[tsv_parts['new_para']]['harmony']
-                    # organise non-doc values as doc:
-                    print("* Universal POS is",
-                          paradata[tsv_parts['new_para']]['upos'],
-                          "and legacy POS is",
-                          paradata[tsv_parts['new_para']]['pos'],
-                          file=outfile)
-                    if harmony == 'both':
-                        print("* suffixes can use " +
-                              "both vowel harmonies", file=outfile)
-                    elif harmony == 'N/A':
-                        print("* no vowel harmony applies to this paradigm",
-                              file=outfile)
-                    else:
-                        print("* suffixes use", harmony, "vowel harmony",
-                              file=outfile)
-                    if 'N/A' not in kotus:
-                        print("* KOTUS paradigm used in",
-                              "their dictionary is", kotus, file=outfile)
-                    else:
[…]
