### 2020-07-28 c04f485
update pip packaging for new version
