+++ src/python/README.md
@@ -1,3 +1,11 @@
 # omorfi-python
 
-Python API for omorfi (the open morphology of Finnish).
+Python API for omorfi (the open morphology of Finnish). Omorfi is a generic
+toolkit for natural language processing of Finnish texts and data. The pip
+packaging allows download and use of some of the language models, however, not
+all parts have python support, e.g. advanced disambiguation.
+
+For more information please see:
+
+* [omorfi on github](https://github.com/flammie/omorfi)
+* [omorfi web site](https://flammie.github.io/omorfi)
