### 2020-07-28 #2184 @flammie
I’ve pushed new release today with pip support and some fixes to the documentation hopefully...
