### 2020-08-21 30576af
revamped homonym definitions
now only real homonyms within same POS, that have separate inflectional
paradigms are considered special. All others can be split outside core
lexicon. Few regressions while smoshing paradings together...
