### 2020-08-28 89720ca
validity fixes
