### 2020-09-02 b0e85f8
fix more regressions and harmony
* s/neljästtä/neljättä/g
* add both harmony class (also for N/A atm)
