### 2020-09-02 663536c
show all test logs at the end
