### 2020-09-02 5f7004f
missing python test part for harmony
