### 2020-09-04 c98230e
optional homonym numbers
fixes #67
