+++ src/test-scripts/lemmas-match-harmony.py
@@ -65,7 +65,7 @@ def main():
             if tsv_parts['harmony'] == 'both':
                 continue
             if tsv_parts['harmony'] == 'back':
-                harmony_regex = '[AOUŌÓÅaouåôąáăâāóōúūû][^äöy]*'
+                harmony_regex = '[AOUŌÓÅaouůåôąáăâāóōúūû][^äöy]*'
             elif tsv_parts['harmony'] == 'front':
                 harmony_regex = '[ÄÖYIEĒÉÍäöyieíüýéőèěęøæõ][^aou]*'
             else:
