### 2020-09-21 b453839
python sort for tsvs for improved stability?
