### 2020-09-21 cc2490b
very basic TUI for adding words
