+++ src/python/new-words.py
@@ -1,0 +1,227 @@
+#!/usr/bin/env python3
+# -*- coding: utf8 -*-
+"""
+Insert list of words to TSV database with guessing of paradigms.
+"""
+
+# Author: Tommi A. Pirinen <flammie@iki.fi> 2009, 2011
+
+#   This program is free software: you can redistribute it and/or modify
+#   it under the terms of the GNU General Public License as published by
+#   the Free Software Foundation, either version 3 of the License, or
+#   (at your option) any later version.
+#
+#   This program is distributed in the hope that it will be useful,
+#   but WITHOUT ANY WARRANTY; without even the implied warranty of
+#   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#   GNU General Public License for more details.
+#
+#   You should have received a copy of the GNU General Public License
+#   along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+import argparse
+import csv
+from sys import exit, stderr
+
+
+def choose_from(options):
+    print("x", "none of the below")
+    for i, opt in enumerate(options):
+        print(i, opt)
+    a = input("x or " + str(range(len(options))) + ": ")
+    while True:
+        try:
+            return options[int(a)]
+        except:
+            if a == 'x':
+                return None
+            else:
+                print(a, "did not work")
+        a = input("once more: ")
+
+def is_back(s):
+    if 'a' in s or 'o' in s or 'u' in s:
+        return True
+    else:
+        return False
+
+def guess_new_para_interactive(wordmap):
+    print("What is:", wordmap)
+    if not wordmap['lemma']:
+        print("No guessing paras without lemmas yet")
+        return 'X_IGNORE'
+    s = wordmap['lemma']
+    para = None
+    if s.endswith('kko'):
+        para = choose_from(['NOUN_UKKO', 'NOUN_LEPAKKO', 'PROPN_UKKO',
+                            'PROPN_LEPAKKO', 'ADJ_KOLKKO', 'ADJ_HUPAKKO'])
+    elif s.endswith('nko'):
+        para = choose_from(['NOUN_RUNKO', 'PROPN_RUNKO', 'ADJ_LENKO'])
+    elif s.endswith('oko'):
+        para = choose_from(['NOUN_RUOKO', 'PROPN_RUOKO', 'NOUN_KOKO',
+                            'PROPN_KOKO'])
+    elif s.endswith('tto'):
+        para = choose_from(['NOUN_HIRTTO', 'PROPN_HIRTTO'])
+    if para:
+        return para
+    if s.endswith('ko'):
+        para = choose_from(['NOUN_PELKO', 'NOUN_VIHKO', 'NOUN_TEKO',
+                            'PROPN_PELKO', 'PROPN_VIHKO', 'PROPN_TEKO',
+                            'ADJ_LAKO'])
+    if para:
+        return para
+    if s.endswith('o'):
+        para = choose_from(['NOUN_TALO', 'PROPN_TALO', 'ADJ_TUMMAHKO'])
+    if para:
+        return para
+    if s.endswith('u'):
+        para = choose_from(['NOUN_ASU', 'NOUN_SEIKKAILU', 'NOUN_KUNGFU',
+                            'PROPN_ASU', 'PROPN_SEIKKAILU', 'PROPN_KUNGFU',
+                            'ADJ_VALKAISTU'])
+    if para:
+        return para
+    if s.endswith('y'):
+        para = choose_from(['NOUN_KÄRRY', 'NOUN_VEHKEILY', 'NOUN_GAY',
+                            'PROPN_KÄRRY', 'PROPN_SPOTIFY', 'PROPN_GAY',
[…]
