### 2020-11-23 6eefc41
update generate-all-wordforms a bit
