### 2020-12-08 52e6330
move some test corpora to subdir
