+++ test/ftb-3.1-test.bash.in
@@ -4,7 +4,7 @@ if test -z $srcdir ; then
     exit 1
 fi
 fsa="@top_builddir@/src/generated/omorfi.describe.hfst"
-cuf="ftb3.1.cutted.freqs"
+cuf="corpora/ftb3.1.cutted.freqs"
 cs="get-covered.bash"
 if test ! -r "$fsa" ; then
     echo Missing $fsa
