+++ docs/generate-mwe.py
@@ -1,0 +1,201 @@
+#!/usr/bin/env python3
+# -*- coding: utf8 -*-
+"""
+This script converts Finnish TSV-formatted lexicon to github wiki
+"""
+
+
+# Author: Tommi A Pirinen <flammie@iki.fi> 2009, 2011
+
+#   This program is free software: you can redistribute it and/or modify
+#   it under the terms of the GNU General Public License as published by
+#   the Free Software Foundation, either version 3 of the License, or
+#   (at your option) any later version.
+#
+#   This program is distributed in the hope that it will be useful,
+#   but WITHOUT ANY WARRANTY; without even the implied warranty of
+#   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#   GNU General Public License for more details.
+#
+#   You should have received a copy of the GNU General Public License
+#   along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+
+import argparse
+import csv
+from sys import stderr
+
+
+def filenamify(s):
+    repls = {'!': 'EXCL', '?': 'QQ', '"': 'DQUO', "'": 'SQUO', '/': 'SLASH',
+             '\\': 'BACKSLASH', '<': 'LEFT', '>': 'RIGHT', ':': 'COLON',
+             ' ': 'SPACE', '|': 'PIPE', '.': 'STOP', ',': 'COMMA',
+             ';': 'SC', '(': 'LBR', ')': 'RBR', '[': 'LSQ',
+             ']': 'RSQ', '{': 'LCURL', '}': 'RCURL', '$': 'DORA',
+             '\t': 'TAB'}
+    for needl, subst in repls.items():
+        s = s.replace(needl, subst)
+    if s[0].upper() in 'ABCDEFGHIJKLMNOPQRSTUVWXYZÅÄÖŠŽ':
+        s = s[0].upper() + '/' + s
+    return s
+
+
+def markdownify(s):
+    repls = {']': '(right square bracket)', '|': '(pipe symbol)',
+             '[': '(left square bracket)'}
+    for needl, subst in repls.items():
+        s = s.replace(needl, subst)
+    return s
+
+
+def wiktify(s):
+    repls = {']': '(right square bracket)', '|': '(pipe symbol)',
+             '[': '(left square bracket)'}
+    for needl, subst in repls.items():
+        s = s.replace(needl, subst)
+    return s
+
+
+def homonymify(s):
+    # ₀₁₂₃₄₅₆₇₈₉
+    if '2' in s:
+        return '₂'
+    elif '3' in s:
+        return '₃'
+    elif '4' in s:
+        return '₄'
+    elif '5' in s:
+        return '₅'
+    elif '6' in s:
+        return '₆'
+    elif '7' in s:
+        return '₇'
+    elif '8' in s:
+        return '₈'
+    elif '9' in s:
+        return '₉'
+    else:
+        return '₁'
+
+
+def stuff2icon(s):
+    if s == 'ORG':
+        return 'org'
+    elif s == 'GEO':
+        return '🌍'
+    elif s == 'FIRST':
+        return '🧑¹'
+    elif s == 'LAST':
+        return '🧑²'
+    elif s == 'FEMALE':
+        return '♀'
+    elif s == 'MALE':
+        return '♂'
+    elif s == 'CURRENCY':
+        return '💱'
+    elif s == 'MEDIA':
+        return '📺'
+    elif s == 'MISC':
+        return '⁈'
+    elif s == 'CULTGRP':
+        return '🎶'
+    elif s == 'LANGUAGE':
+        return '🗪'
+    elif s == 'COUNTRY':
+        return '⚐'
+    elif s == 'MEASURE':
+        return '📏'
+    else:
+        return s
+
+# standard UI stuff
+
+def main():
[…]
