+++ src/python/omorfi-ner.py
@@ -1,0 +1,204 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+"""A command-line interface for omorfi NER parsing."""
+
+# string munging
+from argparse import ArgumentParser, FileType
+# CLI stuff
+import sys
+from sys import stderr, stdin, stdout
+# statistics
+from time import perf_counter, process_time
+# omorfi
+import csv
+from omorfi import Omorfi
+from omorfi.formats.fileformats import next_finer
+
+
+def propntype2ner(s):
+    if s == "Geo":
+        return "LOC"
+    elif s == "First":
+        return "PER"
+    elif s == "Last":
+        return "PER"
+    elif s == "Org":
+        return "ORG"
+    elif s == "Product":
+        return "PRO"
+    elif s == "Cultgrp":
+        return "ORG" # XXX
+    elif s == "Misc":
+        return "ORG" # XXX
+    else:
+        print("PropnType missing:", s)
+        sys.exit(1)
+        return "PRO" # XXX
+
+
+def get_ner_b(analysis):
+    if "PropnType" in analysis.misc:
+        return "B-" + propntype2ner(analysis.misc["PropnType"])
+    else:
+        return "O" # XXX
+
+
+def main():
+    """Invoke a simple CLI analyser."""
+    a = ArgumentParser()
+    a.add_argument('-a', '--analyser', metavar='AFILE', required=True,
+                   help="read analyser model from AFILE")
+    a.add_argument('-i', '--input', metavar="INFILE", type=open,
+                   dest="infile", help="source of analysis data")
+    a.add_argument('-v', '--verbose', action='store_true',
+                   help="print verbosely while processing")
+    a.add_argument('-o', '--output', metavar="OUTFILE", dest="outfile",
+                   help="print output into OUTFILE", type=FileType('w'))
+    a.add_argument('-x', '--statistics', metavar="STATFILE", dest="statfile",
+                   help="print statistics to STATFILE", type=FileType('w'))
+    a.add_argument('-M', '--mwe', metavar="MWEFILE",
+                   help="open mwe master db tsv from MWEFILE")
+    a.add_argument('--debug', action='store_true',
+                   help="print lots of debug info while processing")
+    options = a.parse_args()
+    if options.verbose:
+        print("Printing verbosely")
+    omorfi = Omorfi(options.verbose)
+    if options.analyser:
+        if options.verbose:
+            print("reading analyser model", options.analyser)
+        omorfi.load_analyser(options.analyser)
+    else:
+        print("analyser is needed to mwe", file=stderr)
+        sys.exit(4)
+    mwedata = dict()
+    if options.mwe:
+        with open(options.mwe, "r", newline='') as mwefile:
+            mwe_reader = csv.DictReader(mwefile, delimiter='\t',
+                                        quoting=csv.QUOTE_NONE, escapechar='\\',
+                                        quotechar=None, strict=True)
+            for mwe_parts in mwe_reader:
+                mwedata[mwe_parts['lemma']] = mwe_parts
+    if not options.infile:
+        print("reading from <stdin>")
+        options.infile = stdin
+    if options.verbose:
+        print("analysing", options.infile.name)
+    if not options.outfile:
+        options.outfile = stdout
+    if options.verbose:
+        print("writing to", options.outfile.name)
+    if not options.statfile:
+        options.statfile = stdout
+    # statistics
+    realstart = perf_counter()
+    cpustart = process_time()
+    tokens = 0
+    unknowns = 0
+    sentences = 0
+    eoffed = False
+    previous_ner = list()
+    previous_anal = list()
+    insidemwe = False
+    mwefinished = False
+    while not eoffed:
+        sentplus = next_finer(options.infile)
+        if not sentplus:
[…]
