### 2021-01-13 babab92
apertium needs differnet propn classes too
