+++ src/python/omorfi/formats/apertium_formatter.py
@@ -91,6 +91,7 @@ class ApertiumFormatter(Formatter):
         "np",
         "num",
         "ord",
+        "org",
         "p1",
         "p2",
         "p3",
@@ -291,7 +292,7 @@ class ApertiumFormatter(Formatter):
         "Opl2": "px2pl",
         "ORDINAL": "ord",
         "ORD": "ord",
-        "ORG": "al",
+        "ORG": "org",
         "UNKNOWNPROPN": "al",
         "Osg1": "px1sg",
         "Osg2": "px2sg",
