### 2021-02-25 58b37bb
the best design for simple entry point
so far...
