### 2021-07-12 #2184 @desilinguist
Sorry for the delay on this. I took a look at the updated instructions and while they are certainly better than before, I still think the bar to entry is rather high and this shouldn't really be the case in 2021. I can think of two ways to make this much easier for folks who just want to try out the package:

- Build a conda package. Since conda provides the `c-compiler` and `cxx-compiler` packages, it should be fairly easy to build a self-contained package that installs all the dependencies needed.

- Provide a docker image with all of the dependencies and the package that folks can download and easily try out all the great things it does.

I understand that this submission has been held up for a long time so please consider these only recommendations and not requirements. As far as I am concerned, this submission can now go forward and is ready to be published.

Good work @flammie!
