### 2021-07-15 9f5c8e0
meta for possible future conda package
