### 2021-07-16 #2184 @flammie
Thanks for the feedback, I tried both conda-builder and docker a bit, [a basic Dockerfile](https://github.com/flammie/omorfi/commit/b8c072b3f5e8b4c9b805b153ecc97dbeaa674f7d) was easy to create, but for conda support, one required dependency hfst would need packaging before.
