### 2021-07-17 0120b66
patch up some unfinished refactorings
