### 2021-07-19 c97d0bb
fix refactoring bugs in optionals
