### 2021-07-21 8e9cda2
add anaconda instructions
