+++ requirements.txt
@@ -1,0 +1,1 @@
+hfst == 3.9.1
