+++ src/python/pyproject.toml
@@ -1,0 +1,3 @@
+[build-system]
+requires = ["setuptools", "wheel", "hfst"]
+build-backend = "setuptools.build_meta"
