+++ src/python/setup.cfg
@@ -1,0 +1,30 @@
+[metadata]
+name = omorfi
+version = 0.0.2
+description = Open morphology of Finnish python bindings
+long_description = file: README.md
+keywords = finnish nlp morphology
+license = BSD 3-Clause License
+classifiers =
+    License :: OSI Approved :: GNU General Public License v3 (GPLv3)
+    Programming Language :: Python :: 3
+    Operating System :: OS Independent
+
+[options]
+zip_safe = False
+include_package_data = True
+packages = omorfi
+scripts =
+    omorfi-download.py
+    omorfi-conllu.py
+    omorfi-convert.py
+    omorfi-disamparsulate.py
+    omorfi-factorise.py
+    omorfi-freq-evals.py
+    omorfi-ftb3.py
+    omorfi-segment.py
+    omorfi-unimorph.py
+    omorfi-vislcg.py
+    omorfi-wikitable.py
+install_requires =
+    hfst
