### 2021-10-02 fc8130a
Error handling on model load, fixes #74
